scheduler: Moved integration tests over to scheduler V2
controller: Move cluster client wrapper into utils package


controller/scheduler-v2: Implement main function


controller: WIP switch to scheduler-v2
