+++ controller/jobs.go
@@ -15,7 +15,6 @@ import (
 	"github.com/flynn/flynn/controller/schema"
 	ct "github.com/flynn/flynn/controller/types"
 	"github.com/flynn/flynn/controller/utils"
-	"github.com/flynn/flynn/discoverd/client"
 	"github.com/flynn/flynn/host/resource"
 	"github.com/flynn/flynn/host/types"
 	"github.com/flynn/flynn/pkg/cluster"
@@ -23,7 +22,6 @@ import (
 	"github.com/flynn/flynn/pkg/httphelper"
 	"github.com/flynn/flynn/pkg/postgres"
 	"github.com/flynn/flynn/pkg/random"
-	"github.com/flynn/flynn/pkg/stream"
 )
 
 /* SSE Logger */
@@ -121,61 +119,6 @@ func (r *JobRepo) List(appID string) ([]*ct.Job, error) {
 		jobs = append(jobs, job)
 	}
 	return jobs, nil
-}
-
-type clusterClientWrapper struct {
-	*cluster.Client
-}
-
-func (c clusterClientWrapper) Host(id string) (utils.HostClient, error) {
-	return c.Client.Host(id)
-}
-
-func (c clusterClientWrapper) Hosts() ([]utils.HostClient, error) {
-	hosts, err := c.Client.Hosts()
-	if err != nil {
-		return nil, err
-	}
-	res := make([]utils.HostClient, len(hosts))
-	for i, h := range hosts {
-		res[i] = h
-	}
-	return res, nil
-}
-
-func (c clusterClientWrapper) StreamHostEvents(ch chan *discoverd.Event) (stream.Stream, error) {
-	hostChan := make(chan *discoverd.Event)
-	stream, err := c.Client.StreamHostEvents(hostChan)
-	if err != nil {
-		return nil, err
-	}
-	go func() {
-		for {
-			h, ok := <-hostChan
-			if !ok {
-				close(ch)
-				return
-			}
-			ch <- h
-		}
-	}()
-	return &clusterStream{
-		ch:           ch,
-		parentStream: stream,
-	}, nil
-}
-
-type clusterStream struct {
-	ch           chan *discoverd.Event
-	parentStream stream.Stream
-}
-
-func (cs *clusterStream) Close() error {
-	return cs.parentStream.Close()
-}
-
-func (cs *clusterStream) Err() error {
-	return cs.parentStream.Err()
 }
 
 func (c *controllerAPI) connectHost(ctx context.Context) (utils.HostClient, string, error) {
